# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "JsonObject": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "title": "JsonObject",
        "type": "object"
      },
      "JsonObjectList": {
        "items": {
          "additionalProperties": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "type": "object"
        },
        "title": "JsonObjectList",
        "type": "array"
      },
      "JsonValue": {},
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "request_id",
          "retryable"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "SeasonType": {
        "enum": [
          "Regular Season",
          "Playoffs"
        ],
        "title": "SeasonType",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Typed analytics and shot-quality model API.",
    "title": "NBA Stat Analyzer API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/compare": {
      "get": {
        "operationId": "compare_players_api_v1_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "a",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "A",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "b",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "B",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Compare Players",
        "tags": [
          "league"
        ]
      }
    },
...
    "/api/v1/league/improvers": {
      "get": {
        "operationId": "improvers_api_v1_league_improvers_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "TS_PCT",
              "enum": [
                "TS_PCT",
                "USG_PCT",
                "AST_PCT",
                "REB_PCT",
                "OFF_RATING",
                "DEF_RATING",
                "NET_RATING",
                "PACE"
              ],
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 15,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Improvers",
        "tags": [
          "league"
        ]
      }
    },
    "/api/v1/league/leaders": {
      "get": {
        "operationId": "leaders_api_v1_league_leaders_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "stat",
            "required": false,
            "schema": {
              "default": "PTS",
              "enum": [
                "PTS",
                "REB",
                "AST",
                "STL",
                "BLK",
                "TOV",
                "FGM",
                "FGA",
                "FG_PCT",
                "FG3M",
                "FG3A",
                "FG3_PCT",
                "FTM",
                "FTA",
                "FT_PCT",
                "OREB",
                "DREB",
                "PLUS_MINUS",
                "MIN",
                "TS_PCT",
                "USG_PCT",
                "OFF_RATING",
                "DEF_RATING",
                "NET_RATING",
                "PACE"
              ],
              "title": "Stat",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "per_mode",
            "required": false,
            "schema": {
              "default": "PerGame",
              "enum": [
                "Totals",
                "PerGame",
                "Per36",
                "Per100Possessions"
              ],
              "title": "Per Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "measure",
            "required": false,
            "schema": {
              "default": "Base",
              "enum": [
                "Base",
                "Advanced"
              ],
              "title": "Measure",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 15,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Leaders",
        "tags": [
          "league"
        ]
      }
    },
    "/api/v1/league/low-minutes-efficient": {
      "get": {
        "operationId": "low_minutes_api_v1_league_low_minutes_efficient_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          },
          {
            "in": "query",
            "name": "max_mpg",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 35,
              "minimum": 10,
              "title": "Max Mpg",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 15,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Low Minutes",
        "tags": [
          "league"
        ]
      }
    },
    "/api/v1/league/similar/{player_id}": {
      "get": {
        "operationId": "similar_api_v1_league_similar__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Similar",
        "tags": [
          "league"
        ]
      }
    },
    "/api/v1/league/team-defense": {
      "get": {
        "operationId": "team_defense_api_v1_league_team_defense_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "examples": [
                    "2025-26"
                  ],
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "season_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeasonType",
              "default": "Regular Season"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Team Defense",
        "tags": [
          "league"
        ]
      }
    },
    "/api/v1/meta": {
      "get": {
        "operationId": "meta_api_v1_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Meta",
        "tags": [
          "league"
        ]
      }
    },
...
  }
}